+++ src/main/java/de/retest/recheck/cli/utils/GoldenMasterUtil.java
@@ -1,0 +1,11 @@
+package de.retest.recheck.cli.utils;
+
+public class GoldenMasterUtil {
+
+	public static final String GOLDEN_MASTER_PARAMETER_DESCRIPTION =
+			"Path to a golden master folder. This command may also "
+					+ "be used with a single parameter to print out a test report (deprecated)."; //
+
+	private GoldenMasterUtil() {}
+
+}