+++ src/test/java/de/retest/recheck/cli/subcommands/DiffIT.java
@@ -3,6 +3,7 @@ package de.retest.recheck.cli.subcommands;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.File;
+import java.io.IOException;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -11,12 +12,13 @@ import org.junit.contrib.java.lang.system.SystemOutRule;
 import org.junit.rules.TemporaryFolder;
 
 import de.retest.recheck.RecheckProperties;
+import de.retest.recheck.cli.testutils.GoldenMasterCreator;
 import de.retest.recheck.cli.testutils.ProjectRootFaker;
 import de.retest.recheck.cli.testutils.TestReportCreator;
+import de.retest.recheck.util.FileUtil;
 import picocli.CommandLine;
 
 public class DiffIT {
-
 	@Rule
 	public final TemporaryFolder temp = new TemporaryFolder();
 
@@ -28,22 +30,210 @@ public class DiffIT {
 
 	@Test
 	public void diff_without_argument_should_return_the_usage_message() {
-		final String expected =
-				"Usage: diff [--exclude=<exclude>]... <testReport>\n" + "Display differences of given test report.\n"
-						+ "      <testReport>          Path to a test report file (.report extension). If\n"
-						+ "                              the test report is not in the project directory,\n"
-						+ "                              please specify the absolute path, otherwise a\n"
-						+ "                              relative path is sufficient.\n"
-						+ "      --exclude=<exclude>   Filter to exclude changes from the report. For a\n"
-						+ "                              custom filter, please specify the absolute path.\n"
-						+ "                              For predefined filters, a relative path is\n"
-						+ "                              sufficient. Specify this option multiple times to\n"
-						+ "                              use more than one filter.\n";
+		final String expected = "Usage: diff [--output=<directory>] [--exclude=<exclude>]...\n"
+				+ "            <goldenMasterPath>..." + "Compare two Golden Masters.\n"
+				+ "      <goldenMasterPath>...  Path to a golden master folder. This command may\n"
+				+ "                               also be used with a single parameter to print\n"
+				+ "                               out a test report (deprecated).\n"
+				+ "      --exclude=<exclude>    Filter to exclude changes from the report. For a\n"
+				+ "                               custom filter, please specify the absolute path.\n"
+				+ "                               For predefined filters, a relative path is\n"
+				+ "                               sufficient. Specify this option multiple times\n"
+				+ "                               to use more than one filter.\n"
+				+ "      --output=<directory>   Save differences of Golden Masters to specified\n"
+				+ "                               directory as test report.\n";
+
 		assertThat( new CommandLine( new Diff() ).getUsageMessage() ).isEqualToIgnoringNewLines( expected );
 	}
 
 	@Test
-	public void diff_should_print_differences() throws Exception {
+	public void diff_should_give_proper_error_message_when_given_golden_master_path_does_not_exist() throws Exception {
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File outputDir = temp.newFolder( "somedirectory" );
+
+		final File doesNotExist = temp.newFile( "doesnotexist" );
+		final String[] args = { "--output", outputDir.getAbsolutePath(), doesNotExist.getAbsolutePath(), "does/exist" };
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		assertThat( systemOutRule.getLog() ).contains( "Could not load Golden Master from file '"
+				+ FileUtil.canonicalFileQuietly( new File( doesNotExist, RecheckProperties.DEFAULT_XML_FILE_NAME ) )
+				+ "'." );
+	}
+
+	@Test
+	public void diff_should_provide_proper_error_message_when_provided_file_not_a_golden_master() throws IOException {
+		temp.newFolder( "path" );
+		temp.newFolder( "path", "notagoldenmaster" );
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File notAGoldenMaster = temp.newFolder( "path", "notagoldenmaster", "a" );
+		final File aGoldenMaster = temp.newFolder( "someotherpath", "goldenmaster", "b" );
+
+		final File outputDir = temp.newFolder( "somedirectory" );
+
+		GoldenMasterCreator.createErroneousGoldenMasterFile( notAGoldenMaster );
+		GoldenMasterCreator.createGoldenMasterFile( aGoldenMaster, true );
+
+		final File expectedPathErroneousFile =
+				new File( FileUtil.canonicalFileQuietly( notAGoldenMaster ), RecheckProperties.DEFAULT_XML_FILE_NAME );
+
+		final String[] args = { "--output", outputDir.getAbsolutePath(), notAGoldenMaster.getAbsolutePath(),
+				aGoldenMaster.getAbsolutePath() };
+
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		assertThat( systemOutRule.getLog() ).containsSubsequence(
+				"Could not load Golden Master from file '" + expectedPathErroneousFile.getAbsolutePath() + "'." );
+
+	}
+
+	@Test
+	public void diff_should_print_differences_of_golden_masters() throws IOException {
+		temp.newFolder( "path" );
+		temp.newFolder( "path", "goldenmaster" );
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File outputDir = temp.newFolder( "somedirectory" );
+
+		final File gm1 = temp.newFolder( "path", "goldenmaster", "a" );
+		final File gm2 = temp.newFolder( "someotherpath", "goldenmaster", "b" );
+
+		GoldenMasterCreator.createGoldenMasterFile( gm1, true );
+		GoldenMasterCreator.createGoldenMasterFile( gm2, false );
+
+		final String[] args = { "--output", outputDir.getAbsolutePath(), gm1.getAbsolutePath(), gm2.getAbsolutePath() };
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		final String expected = "Comparison of Golden Masters resulted in:\n" //
+				+ "\tMetadata Differences:\n" //
+				+ "\t	some.driver: expected=\"driverA\", actual=\"driverB\"\n" //
+				+ "\tbaz [changed text] at 'foo[1]/bar[1]/baz[1]':\n" //
+				+ "\t	was deleted\n" //
+				+ "\tbaz [original text] at 'foo[1]/bar[1]/baz[1]':\n" //
+				+ "\t	was inserted\n";
+
+		assertThat( systemOutRule.getLog() ).containsSubsequence( expected );
+	}
+
+	@Test
+	public void diff_should_not_persist_testreport_if_no_output_option() throws IOException {
+		temp.newFolder( "path" );
+		temp.newFolder( "path", "goldenmaster" );
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File gm1 = temp.newFolder( "path", "goldenmaster", "a" );
+		final File gm2 = temp.newFolder( "someotherpath", "goldenmaster", "b" );
+
+		GoldenMasterCreator.createGoldenMasterFile( gm1, true );
+		GoldenMasterCreator.createGoldenMasterFile( gm2, false );
+
+		final String[] args = { gm1.getAbsolutePath(), gm2.getAbsolutePath() };
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		assertThat( systemOutRule.getLog() )
+				.containsSubsequence( "Overall, recheck found 2 difference(s) when checking 2 element(s).\n" );
+	}
+
+	@Test
+	public void diff_should_persist_testreport_if_output_option() throws IOException {
+		temp.newFolder( "path" );
+		temp.newFolder( "path", "goldenmaster" );
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File outputDir = temp.newFolder( "somedirectory" );
+
+		final File gm1 = temp.newFolder( "path", "goldenmaster", "a" );
+		final File gm2 = temp.newFolder( "someotherpath", "goldenmaster", "b" );
+
+		GoldenMasterCreator.createGoldenMasterFile( gm1, true );
+		GoldenMasterCreator.createGoldenMasterFile( gm2, false );
+
+		final File outputFile = new File( outputDir, RecheckProperties.AGGREGATED_TEST_REPORT_FILE_NAME );
+
+		final String[] args = { "--output", outputDir.getAbsolutePath(), gm1.getAbsolutePath(), gm2.getAbsolutePath() };
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		assertThat( outputFile ).exists();
+	}
+
+	@Test
+	public void diff_should_apply_ignore_filter_with_exclude_option() throws IOException {
+		temp.newFolder( "path" );
+		temp.newFolder( "path", "goldenmaster" );
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File outputDir = temp.newFolder( "somedirectory" );
+
+		final File gm1 = temp.newFolder( "path", "goldenmaster", "a" );
+		final File gm2 = temp.newFolder( "someotherpath", "goldenmaster", "b" );
+
+		GoldenMasterCreator.createGoldenMasterFile( gm1, true );
+		GoldenMasterCreator.createGoldenMasterFile( gm2, false );
+
+		final File outputFile = new File( outputDir, RecheckProperties.AGGREGATED_TEST_REPORT_FILE_NAME );
+
+		final String[] args = { gm1.getAbsolutePath(), gm2.getAbsolutePath(), "--output", outputDir.toString() };
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		assertThat( outputFile ).exists();
+	}
+
+	@Test
+	public void diff_should_print_warning_when_exclude_option_with_invalid_filters() throws IOException {
+		temp.newFolder( "path" );
+		temp.newFolder( "path", "goldenmaster" );
+		temp.newFolder( "someotherpath" );
+		temp.newFolder( "someotherpath", "goldenmaster" );
+
+		final File outputDir = temp.newFolder( "somedirectory" );
+
+		final File gm1 = temp.newFolder( "path", "goldenmaster", "a" );
+		final File gm2 = temp.newFolder( "someotherpath", "goldenmaster", "b" );
+
+		GoldenMasterCreator.createGoldenMasterFile( gm1, true );
+		GoldenMasterCreator.createGoldenMasterFile( gm2, false );
+
+		final File outputFile = new File( outputDir, RecheckProperties.AGGREGATED_TEST_REPORT_FILE_NAME );
+
+		final String[] args = { "--exclude", "sty-attributes.filter", "--exclude", "invisib.filter", "--output",
+				outputDir.getAbsolutePath(), gm1.getAbsolutePath(), gm2.getAbsolutePath() };
+		final Diff cut = new Diff();
+		new CommandLine( cut ).parseArgs( args );
+
+		cut.run();
+
+		assertThat( systemOutRule.getLog() )
+				.endsWith( "The invalid filter files are: sty-attributes.filter, invisib.filter\n" );
+	}
+
+	// remove this test in next version when old diff functionality is provided only under new command (see RET-1956)
+	@Test
+	public void diff_should_print_diff_of_test_report_if_used_with_single_parameter() throws IOException {
 		ProjectRootFaker.fakeProjectRoot( temp.getRoot().toPath() );
 		final String[] args =
 				{ TestReportCreator.createTestReportFileWithDiffs( temp, "diff_should_print_differences" ) };
@@ -51,6 +241,7 @@ public class DiffIT {
 		new CommandLine( cut ).parseArgs( args );
 
 		cut.run();
+
 		final String expected = "Suite 'diff_should_print_differences' has 3 difference(s) in 1 test(s):\n" //
 				+ "\tTest 'test' has 3 difference(s) in 3 state(s):\n" //
 				+ "\tcheck resulted in:\n" //
@@ -66,29 +257,30 @@ public class DiffIT {
 		assertThat( systemOutRule.getLog() ).contains( expected );
 	}
 
+	// remove this test in next version when old diff functionality is provided only under new command (see RET-1956)
 	@Test
-	public void diff_should_give_proper_error_message_when_given_test_report_is_not_a_test_report() throws Exception {
-		final File notATestReport = temp.newFile();
-		final String[] args = { notATestReport.getAbsolutePath() };
+	public void diff_should_print_diff_of_test_report_if_used_with_single_parameter_and_exclude_option()
+			throws IOException {
+		ProjectRootFaker.fakeProjectRoot( temp.getRoot().toPath() );
+		final String[] args = { "--exclude", "invisible-attributes.filter", "--exclude", "positioning.filter",
+				TestReportCreator.createTestReportFileWithDiffs( temp ) };
 		final Diff cut = new Diff();
 		new CommandLine( cut ).parseArgs( args );
 
 		cut.run();
 
-		assertThat( systemOutRule.getLog() ).contains(
-				"The given file is not a test report. Please only pass files using the '.report' extension." );
+		final String expected = "\tTest 'test' has 3 difference(s) in 3 state(s):\n" //
+				+ "\tcheck resulted in:\n" //
+				+ "\t	baz [original text] at 'foo[1]/bar[1]/baz[1]':\n" //
+				+ "\t		text: expected=\"original text\", actual=\"changed text\"\n" //
+				+ "\tcheck resulted in:\n" //
+				+ "\t	baz [original text] at 'foo[1]/bar[1]/baz[1]':\n" //
+				+ "\t		was deleted\n" //
+				+ "\tcheck resulted in:\n" //
+				+ "\t	baz [changed text] at 'foo[1]/bar[1]/baz[1]':\n" //
+				+ "\t		was inserted";
+
+		assertThat( systemOutRule.getLog() ).contains( expected );
 	}
 
-	@Test
-	public void diff_should_give_proper_error_message_when_given_test_report_does_not_exist() throws Exception {
-		final File doesNotExist = new File( "/does/not/exist" + RecheckProperties.TEST_REPORT_FILE_EXTENSION );
-		final String[] args = { doesNotExist.getAbsolutePath() };
-		final Diff cut = new Diff();
-		new CommandLine( cut ).parseArgs( args );
-
-		cut.run();
-
-		assertThat( systemOutRule.getLog() ).contains( "The given file report '" + doesNotExist.getAbsolutePath()
-				+ "' does not exist. Please check the given file path." );
-	}
 }